tdf#161909: WIP - notes for NSS profile path
I'm trying to store the sCertDir string from here
https://git.libreoffice.org/core/+/f60a6fa11d8ef5227abec6d59b359ea16bc1f795/xmlsecurity/source/xmlsec/nss/nssinitializer.cxx#374
so I can recall it later on. It contains the path to the Mozilla NSS profile. E.g.:
/home/user/.thunderbird/foobar/

I need to store that path so I can use it here:
https://git.libreoffice.org/core/+/f60a6fa11d8ef5227abec6d59b359ea16bc1f795/xmlsecurity/source/xmlsec/nss/securityenvironment_nssimpl.cxx#102
Currently securityenvironment_nssimpl.cxx calls PK11_GetTokenName(pSlot) which dumbly returns NSS Certificate DB when NSS is being used. And that string is then being displayed in the NSS password dialog.
But instead the Mozilla/NSS profile path should be used there. So the password dialog can tell the user for which Mozilla profile is being used. Else the user won't know which password to enter.

Can I somehow use that XComponentContext to achieve this?

Or should I maybe store that path in a static class variable or global variable?
I know, that sounds crazy ;-)
But NSS_InitReadWrite() also seems to store it's result in some global state. So it might make sense to do the same with the corresponding path information.

Change-Id: I53a3bab12c5b8f0476190826f6c9947d1a2a31b5