tdf#161909: enable using only GPG explicitly in chooseCertificatesImpl
For ODF encryption only GPG keys are allowed. And key selection for this
currently only works correct, because all X.509 implementations (nss and mscrypt)
give empty results when asked for public keys (encryption).

This commit is a first step to make the request for GPG keys explicit.
Other code still needs to be changed to make use of this new function
argument.

Change-Id: Ia6b8ecb901c0f54a9a58b1bd0efaa7154ffbd285